gigasecond: more work

+++ test/Exercism.Analyzers.CSharp.IntegrationTests/Solutions/Gigasecond/AddSeconds/DigitsWithoutSeparator/ArgumentIsConstantInMethod/Gigasecond.cs
@@ -4,7 +4,7 @@ public static class Gigasecond
 {
     public static DateTime Add(DateTime birthDate)
     {
-        const double seconds = 1000000000;
+        const int seconds = 1000000000;
         return birthDate.AddSeconds(seconds);
     }
 }
