+++ test/Exercism.Analyzers.CSharp.IntegrationTests/Solutions/Gigasecond/AddSeconds/DigitsWithoutSeparator/ArgumentIsStaticField/Gigasecond.cs
@@ -2,7 +2,7 @@ using System;
 
 public static class Gigasecond
 {
-    private static double Seconds = 1000000000;
+    private static int Seconds = 1000000000;
     
     public static DateTime Add(DateTime birthDate)
     {
